update process study of each student update record process student

+++ src/main/java/com/drawproject/dev/dto/ProgressDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class ProgressDTO {
 
-    private int progress;
+    private Long progress;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private String status;
